Refactor escort feature specs
We now use the feature helpers to streamline the process of
filling in escort record forms.
